requirements.txt
pandas
langsmith
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
streamlit
faiss-cpu
python-dotenv
bs4